# srve/documentation/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "MSPR dev : Pharmacie database",
      "description": "API for CRUD operation on pharmacie database"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "tags": [
      {
        "name": "Pharmacie-api",
        "description": "API for CRUD operation on pharmacie database"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/Pharmacies": {
        "post": {
          "tags": [
            "Pharmacie"
          ],
          "description": "Create new pharmacie",
          "parameters": [
            {
              "name": "pharmacie",
              "in": "body",
              "description": "Pharmacie that we want to create",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Pharmacie successfuly added !",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          }
        },
        "get": {
          "tags": [
            "Pharmacie"
          ],
          "summary": "Get all pharmacie in system",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          }
        }
      },
      "/Pharmacie/{pharmacieId}": {
        "parameters": [
          {
            "name": "pharmacieId",
            "in": "path",
            "required": true,
            "description": "ID of pharmacie that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Pharmacie"
          ],
          "summary": "Get pharmacie with given ID",
          "responses": {
            "200": {
              "description": "pharmacie found",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete pharmcacie with given ID",
          "tags": [
            "Pharmacie"
          ],
          "responses": {
            "200": {
              "description": "Pharmacie deleted",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          }
        },
        "put": {
          "summary": "Update pharmacie with given ID",
          "tags": [
            "Pharmacie"
          ],
          "parameters": [
            {
              "name": "pharmacie",
              "in": "body",
              "description": "Pharmacie with new values of properties",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Pharmacie updated",
              "schema": {
                "$ref": "#/definitions/Pharmacie"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Pharmacie": {
        "required": [
          "name",
          "_id"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "uniqueItems": true
          },
          "name": {
            "type": "string",
            "uniqueItems": false
          },
          "address": {
            "type": "string"
          },
          "lattitude": {
            "type": "float"
          },
          "longitude": {
            "type": "float"
          },
          "turnover": {
            "type": "float"
          },
          "traningNeed": {
            "type": "string"
          }
        }
      },
      "Pharmacie": {
        "type": "array",
        "$ref": "#/definitions/Pharmacie"
      }
    }
  }